[ADD] base, l10n_pe: Adding toponyms data
Adding States and Cities for Peru localization.

[REM] l10n_pe: Removing toponyms as data

The toponyms are moved to l10n_pe_base instead of adding them in l10n_pe
and depending on account, because the new structure is:

l10n_pe: all the fields or data that will be needed for account
l10n_pe_base: is an app that will be useful for peru and don't have
anything to be with account data
l10n_pe_edi: everything related to the localization and electronic
invoicing

[ADD] l10n_pe: adding missing taxes

Adding a new set of taxes to cover the cases that we are missing, and
adding the codes to identify the taxes.

Based on the information we find in http://cpe.sunat.gob.pe/sites/default/files/inline-files/AjustesValidacionesCPEv20180726.zip
We are adding the next data to create the taxes.

[ADD] l10n_pe: add new taxes according to catalog 05

Based on the update of 24/06/2019 from http://cpe.sunat.gob.pe/node/88
we are validating the catalog 5 with the most relevant taxes from
catalog 5.

- IVAP: Tax applied for rice at 4%, it does not require IGV applied to
this product.
- ISC: Selective tax for consumers, it is applied to an specific group
of products, usually are products that are nocive to the health of the
consumer an are meant to drive the consumer away from this product.
- ICBPER: Special tax that will be charged to plastic bags starting in
august of 2019.
- EXO: Exonerated taxes are those that depends on special conditions to
allow the IGV(base tax) to don't be applied.
- INA: Unaffected taxes are for those products that will never be taxed
with IGV or any tax.

[ADD] add l10n_pe_base

This module will have all the extra information related to Peru, the vat
validation, the districts(a special level of address data in peru), the
res.city and the template for country, where the special configuration for peru
will be displayed if the company is from Peru.

[ADD] l10n_pe: Add tags and groups to the taxes

Adding the tags that correspond to the "tax code identifier" and
the groups to every existing tax.

[ADD] l10n_pe: Adding perception's data

Adding perception's data as taxes, this data is part of the
catalogs 22 and 53 from SUNAT.
We can find this information in the file "Validation rules, 2.0 and 2.1"
updated in 24/06/2019 from http://cpe.sunat.gob.pe/node/88

[ADD] l10n_pe: Adding detraction's data

Adding goods and services susceptible to detractions as taxes,
and it's codes as tags, this data is part of the catalog 54 from SUNAT.
We can find this information in the file "Validation rules, 2.0 and 2.1"
updated in 24/06/2019 from http://cpe.sunat.gob.pe/node/88

[FIX] base_vat, l10n_pe_base: Fixing Peruvian VAT validations

Adding "DNI", "Alien Registration Card", "Passport", "Diplomatic Identity
Card", "Identity document of the country of residence",
"Tax Identification Number - TIN", "Identification Number - IN" and
"Andean Immigration Card (TAM, Spanish acronym)" validations.
And improving how the "RUC" is been validated.
This is part of the catalog 06 from SUNAT.
We can find this information in the file "Validation rules, 2.0 and 2.1"
updated in 24/06/2019 from http://cpe.sunat.gob.pe/node/88
Or in the codes catalog:
http://cpe.sunat.gob.pe/sites/default/files/inline-files/anexoV-340-2017.pdf

[FIX] l10n_pe: fix errors on data creation

There is no need for tags in taxes for v13.