[FIX] sale: traceback when choosing a product with warning as order.line
If a user chooses a product that has a blocking error message, the product
is set to False. This is good, but we do not want the following to
happen:

When a blocking message is configured on a product, we force the removal
of the `product_id` on the SO line. This causes a problem in method
`name_get()` and name of the product with sudo fails to search for it.
Only work with admin user

We move the `product_id` removal at the end of the onchange.