[PATCH] base: removed login_key constraint to avoid errors on log
When we execute `-u all` to update all modules, Odoo loads the
constraints as it goes through the modules, without taking into account
that the constraints might be removed or updated in a child module.

In this case the `base` module adds a constraint to make the `login`
field unique. Later, the `website` module replaces this constraint by
making unique the combination of `login` and `website_id`, to allow
users to register in the different websites.

The first time, and while there are no users registered on multiple
websites, this doesn't raise any issues. After any users registers in
multiple websites, when we try to run `-u all` again, `base` will try to
recreate the constraint with its original definition, but it will fail
as there are already multiple users with the same `login` (but on
different websites).

This patch is removing the original definition of the constraint,
leaving the work on the `website` module to create it with the new
definition, and avoid the error when running `-u all`.