[FIX] product: Price on pricelist with pricelists inside of it
When a you have a global pricelist and inside of it, you have an
item 'applied on' 'Global' and a pricelist 'A' with product A,
and another item also with 'applied on' 'Global' with  a pricelist 'B'
with the product B.
And you make a sale order with the global pricelist and you add both
products, the only product that is going to have the price set with the
pricelist B or A it depends which item is first.
With this change it will check in all items of the global pricelist
until it find the price or it will set the price on the product.

Also, adding the field `active` on the model `product.pricelist.item`.