[ADD] product: New Key on the method `_compute_price_rule`
Creating a new key 'rule_%i' (%i == product.id) to send the rule that
give us the rule where we obtain the price.

For example: If we have a priceslit "A" with an item (item_1) that have
the price set (supplierinfo or fixed), and we have a pricelist "B" with
the pricelist "A" set as item (item_2).
The suitable rule here will be "item_2" but we want to know the exact
rule that it gave us the price which was the "item_1".

So in this new key, we are returning as tuple the rule "item_1" and
a 0 (to avoid problems on others methods were this method is called).
But if we don't have this example and we only have the pricelist "A"
we are returning the suitable rule which in this case will be the
"item_1".