[FIX] ir.attachment: Use another cursor for the vacuum process
Main
-

Use another cursor for the vacuum process and release it on every cleaning
process

Explanation
-
And yes :X There are cases where a Garbage Collection is to be done on a huge
amount of files. And while doing so other services/crons could be hit by this
GC process and could get stuck with a locking transaction and could get stuck
with a locked transaction. One example of this process is fetchmail.