[REF] sale: Added context value to prevent auto invoice
In a previous development to a client, the payment acquirers have a field
to specify the invoice policy.

Since we are not using the default invoice policy (the one from the config), the original
validation only compares the config parameter, so even if we change the policy
to delivered quantities the invoice will be created in case the default policy has auto invoice checked.