[IMP] point_of_sale: split create payment move entry logic
Odoo's default behaviour is to create a journal entry for the invoice
payment ('entry'), splitting the logic used to create this entry into
a new function would allow l10n modules to alter this behaviour.

As an example of the use case would be Mexico EDI, where a refund
('out_refund') needs to be generated additional to the customer invoice
for correct tax purposes after the POS session has been closed.