[FIX] account: removes limiting `if tax_ids` condition
In odoo the function `map_tax` in account.fiscal.position returns
an empty recordset of account.tax if it receives an empty recordset
as input.

Some extensions of odoo might change this behavior and return a non-
empty recordset even with an empty input. For example map_tax might
return `tax_ids.tax_dest_id` which have `is_fee = True`.

This commit removes the validation `if tax_ids` so that `map_tax`
is always called. In vanilla Odoo this just adds a noop because
it would call a method which loops over zero elements and returns.