[IMP] web: export company selector
Right now it is not possible to inherit the company selector widget.
In order to perform developments that modify the behavior of the
widget, it is necessary to create a new one instead of just inheriting
it, which is not a good practice. This commit makes it possible to
inherit the company selector widget.