[REF] website_event: Refresh recaptcha on submit event registration form.
When you want to register for an event, after selecting the number of tickets
it opens a modal with a form to fill, on this moment it generates a recaptcha
token, but if you take more than 2 minutes to submit the form it will return
an error of timeout recaptcha, which is unavoidable for cases of wanting to
register a great quantity of tickets at the same time, because it takes a long
time to fill the form. To avoid this issue, we refresh the token on the submit.