[IMP] web: Allow setting Char Field as Field's Domain
Allow setting a Char Field as the domain for another field, this will
avoid having to create a `many2many` computed field to have the allowed
registers of another domain field like `[("id", "=",
many2many_allowed_field)]` because that was messing with the performance
when the allowed records was a large amount of them.

Now with this, we only have to create a computed char field with the
dynamic domain as a JSON, and set it like:
`domain="allowed_domain_field"`.

This was refactored in v17 and isn't necessary https://github.com/odoo/odoo/commit/3fac03b7