[REF] hr_expense: Create Vendor Bills in both types of expenses
We need to maintain the Vendor Bills created from the Expenses
Reports as drafts because they will have a different approval
process.

This commit contains changes for the issues 2936 [1] and 3038 [2]
in the IRC project.

It's the forward-port of the PR 606 [3], but part of the code wasn't
needed because it was refactored in Odoo migration [4].

[1]: https://gitlab.com/ircanada/ircodoo/-/issues/2936
[2]: https://gitlab.com/ircanada/ircodoo/-/issues/3038
[3]: https://github.com/Vauxoo/odoo/pull/606
[4]: https://github.com/odoo/odoo/commit/f79ff3fb